+++ murrelet_livecode/src/lib.rs
@@ -3,4 +3,5 @@ pub mod boop;
 pub mod expr;
 pub mod livecode;
 pub mod nestedit;
+pub mod state;
 pub mod unitcells;
